add login, logout and restrictions

+++ base/views.py
@@ -1,14 +1,39 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.contrib import messages
+from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
 # Create your views here.
 
-# rooms = [
-#     {'id': 1, 'name':'Lets learn python' },
-#     {'id': 2, 'name':'Design with me' },
-#     {'id': 3, 'name':'Frontend developers' },
-# ]
+def loginPage(request):                     # there is a built in function called login zato se ova fun zove drugacije
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist')  # lose sto se tice securitya
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)            # new session to db and browser
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password are incorrect')
+
+    context = {}
+    return render(request, 'base/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -23,15 +48,11 @@ def home(request):
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
-    # room = None
-    # for i in rooms:
-    #     if i['id'] == int(pk):
-    #         room = i
-    #         break
     room = Room.objects.get(id=pk)
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')    # redirect if not logged in
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -43,10 +64,13 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('Not your room')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -57,8 +81,13 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('Not your room')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
